refactor: updateRaw is now asRaw
