+++ packages/dataparcels/.size-limit
@@ -25,7 +25,7 @@
     },
     {
         limit: "102 B",
-        path: "updateRaw.js"
+        path: "asRaw.js"
     },
     {
         limit: "15 KB",
