+++ packages/dataparcels/asRaw.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('./lib/parcelData/asRaw.js').default;
