+++ packages/dataparcels/src/validation/__test__/validation-test.js
@@ -2,7 +2,7 @@
 import validation from '../validation';
 
 test('validation should use a raw updater, so it will work in beforeChange', () => {
-    expect(validation({})._updateRaw).toBe(true);
+    expect(validation({})._asRaw).toBe(true);
 });
