--- packages/dataparcels/updateRaw.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-module.exports = require('./lib/parcelData/updateRaw.js').default;
