+++ packages/react-dataparcels/asRaw.js
@@ -1,0 +1,2 @@
+/* eslint-disable */
+module.exports = require('dataparcels/asRaw.js');
