+++ packages/react-dataparcels/src/ParcelBoundaryDeprecated.jsx
@@ -7,7 +7,7 @@ import type {ParcelValueUpdater} from 'dataparcels';
 
 import React from 'react';
 import Parcel from 'dataparcels';
-import updateRaw from 'dataparcels/updateRaw';
+import asRaw from 'dataparcels/asRaw';
 
 import ParcelBoundaryControl from './ParcelBoundaryControlDeprecated';
 import ParcelBoundaryEquals from './util/ParcelBoundaryEquals';
@@ -285,7 +285,7 @@ export default class ParcelBoundary extends React.Component<Props, State> { /* e
             })
             ._changeAndReturn((parcel) => parcel
                 .modifyDown(prevData
-                    ? updateRaw(() => prevData)
+                    ? asRaw(() => prevData)
                     : identity()
                 )
                 .pipe(ApplyModifyBeforeUpdate(modifyBeforeUpdate))
