+++ packages/react-dataparcels/src/util/pipeWithFakePrevParcel.js
@@ -3,7 +3,7 @@
 import type {ParcelUpdater} from 'dataparcels';
 
 import Parcel from 'dataparcels';
-import updateRaw from 'dataparcels/updateRaw';
+import asRaw from 'dataparcels/asRaw';
 
 export default (prevParcel: ?Parcel, ...pipe: ParcelUpdater[]) => (parcel: Parcel): Parcel => {
     let [changedParcel] = parcel._changeAndReturn((parcel) => {
@@ -11,7 +11,7 @@ export default (prevParcel: ?Parcel, ...pipe: ParcelUpdater[]) => (parcel: Parce
         let setPrevParcel;
         if(prevParcel) {
             let prevParcelData = prevParcel.data;
-            setPrevParcel = updateRaw(() => prevParcelData);
+            setPrevParcel = asRaw(() => prevParcelData);
         }
 
         return parcel
