--- packages/react-dataparcels/updateRaw.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-module.exports = require('dataparcels/updateRaw.js');
